updated and added components.hpp
